+++ frontend/src/repositories/todoRepository/createTodoRepository.ts
@@ -11,8 +11,8 @@ export function createTodoRepository(apiClient: ApiClient): TodoRepository {
     create: (input: InputTodo) =>
       apiClient.post<CreateTodoResponse>(path, input, (_data): _data is Todo => true /* isTodo */),
     update: (id: TodoId, input: UpdateTodo) => apiClient.patch(`${path}/${id}`, input),
-    remove: (id: TodoId) => apiClient.remove(`${path}/${id}`),
-    removeAll: (ids: TodoId[]) =>
+    delete: (id: TodoId) => apiClient.delete(`${path}/${id}`),
+    deleteAll: (ids: TodoId[]) =>
       apiClient.put<RemoveAllTodoResponse>(path, { ids: ids }, (_data): _data is Todo[] => true /* isTodoArray */)
   };
 }